# dist/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Discord_bis API",
    "description": "Mini projet webServices API",
    "termsOfService": "http://swagger.io/terms/"
  },
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "Create an user",
        "operationId": "Http://localhost:3001/users",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "User Parameters",
            "in": "body",
            "description": "Create an user in our database",
            "required": true,
            "schema": {
              "$ref": "#/components/Users"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Create response",
            "schema": {
              "$ref": "#/components/Users"
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/users/{user_id}": {
      "delete": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Users"
        ],
        "description": "Delete an user",
        "operationId": "Delete user by Id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "description": "ID of user to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "User successfully deleted"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Access token does not have the required scope"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Returns the status of an user (just the username and email)",
        "operationId": "Find by Id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "description": "ID of user to fetch",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "response",
            "schema": {
              "$ref": "#/components/UsersDTO"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/channels": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Channels"
        ],
        "description": "Create a channel",
        "operationId": "Http://localhost:3001/channels",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Channel Parameters",
            "in": "body",
            "description": "Create a channel in a server in our database",
            "required": true,
            "schema": {
              "$ref": "#/components/Channels"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "create response",
            "schema": {
              "$ref": "#/components/Channels"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Access token does not have the required scope"
          }
        }
      }
    },
    "/channels/{channel_id}": {
      "delete": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Channels"
        ],
        "description": "Delete a channel",
        "operationId": "Delete channel by Id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "channel_id",
            "description": "ID of channel to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Delete response",
            "schema": {
              "$ref": "#/components/Channels"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Access token does not have the required scope"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Channels"
        ],
        "description": "Returns the status of a channel",
        "operationId": "Find channel by Id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "channel_id",
            "description": "ID of channel to fetch",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "response",
            "schema": {
              "$ref": "#/components/Channels"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Access token does not have the required scope"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Channels"
        ],
        "description": "Update a channel",
        "operationId": "Update channel by Id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "channel_id",
            "description": "ID of channel to update",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "channel update",
            "description": "channel to update",
            "required": true,
            "schema": {
              "$ref": "#/components/Channel_update"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Channel successfully updated",
            "schema": {
              "$ref": "#/components/Channels"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Access token does not have the required scope"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/servers": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Servers"
        ],
        "description": "Create a server",
        "operationId": "Http://localhost:3001/servers",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Server Parameters",
            "in": "body",
            "description": "Create a server in our database",
            "required": true,
            "schema": {
              "$ref": "#/components/Servers"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server successfully created",
            "schema": {
              "$ref": "#/components/Servers"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Access token does not have the required scope"
          }
        }
      }
    },
    "/servers/{server_id}": {
      "delete": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Servers"
        ],
        "description": "Delete a server",
        "operationId": "Delete server by Id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "description": "ID of server to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Server successfully deleted",
            "schema": {
              "$ref": "#/components/Servers"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Access token does not have the required scope"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Servers"
        ],
        "description": "Returns the status of a server",
        "operationId": "Find server by Id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "description": "ID of server to fetch",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "response",
            "schema": {
              "$ref": "#/components/Servers"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Access token does not have the required scope"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Servers"
        ],
        "description": "Update a server",
        "operationId": "Update server by Id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "description": "ID of server to update",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "modication content",
            "description": "server update",
            "required": false,
            "schema": {
              "$ref": "#/components/Server_update"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server successfully update",
            "schema": {
              "$ref": "#/components/Servers"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Access token does not have the required scope"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/messages": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Messages"
        ],
        "description": "Create a message",
        "operationId": "Create a message",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Message Parameters",
            "in": "body",
            "description": "Create a message in our database",
            "required": true,
            "schema": {
              "$ref": "#/components/Messages"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message successfully created",
            "schema": {
              "$ref": "#/components/Messages"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Access token does not have the required scope"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
...
}